requirements.txt
# Database
DATABASE_URL= "mongodb+srv://vkakarla_db_user:evPGYfRQ2yA9Edfb@cluster0.wcn4zgb.mongodb.net/?appName=Cluster0"

# Auth
JWT_SECRET=your_secret_key
JWT_EXPIRATION=7d

# Stripe
STRIPE_SECRET_KEY=sk_test_...
STRIPE_PUBLISHABLE_KEY=pk_test_...

# AWS S3
AWS_ACCESS_KEY_ID=your_key
AWS_SECRET_ACCESS_KEY=your_secret
AWS_S3_BUCKET=neighborly-tools

# Google Maps
GOOGLE_MAPS_API_KEY=your_key

# Twilio
TWILIO_ACCOUNT_SID=your_sid
TWILIO_AUTH_TOKEN=your_token
TWILIO_PHONE_NUMBER=+1234567890

# Firebase
FIREBASE_PROJECT_ID=neighborly-app
FIREBASE_PRIVATE_KEY=...


fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==7.8.0
twilio==8.10.0
requests==2.31.0
EOF